Refactor utils, fix #7 (PR #9)
* Break out config utils into separate module #7 #8

* Break out misc util functions into separate module #7 #8

* Rename remaining utils to email_utils #7 #8

* Refactor utils tests #7
