+++ pdfebc_core/compress.py
@@ -11,7 +11,7 @@ compress PDF files.
 import os
 import sys
 import subprocess
-from . import utils
+from .misc_utils import if_callable_call_with_formatted_string
 
 BYTES_PER_MEGABYTE = 1024**2
 FILE_SIZE_LOWER_LIMIT = BYTES_PER_MEGABYTE
@@ -65,11 +65,11 @@ def compress_pdf(filepath, output_path, ghostscript_binary, status_callback=None
     try:
         file_size = os.stat(filepath).st_size
         if file_size < FILE_SIZE_LOWER_LIMIT:
-            utils.if_callable_call_with_formatted_string(status_callback, NOT_COMPRESSING,
-                                                         filepath, file_size, FILE_SIZE_LOWER_LIMIT)
+            if_callable_call_with_formatted_string(status_callback, NOT_COMPRESSING,
+                                                   filepath, file_size, FILE_SIZE_LOWER_LIMIT)
             process = subprocess.Popen(['cp', filepath, output_path])
         else:
-            utils.if_callable_call_with_formatted_string(status_callback, COMPRESSING, filepath)
+            if_callable_call_with_formatted_string(status_callback, COMPRESSING, filepath)
             process = subprocess.Popen(
                 [ghostscript_binary, "-sDEVICE=pdfwrite",
                  "-dCompatabilityLevel=1.4", "-dPDFSETTINGS=/ebook",
@@ -77,11 +77,11 @@ def compress_pdf(filepath, output_path, ghostscript_binary, status_callback=None
                  "-sOutputFile=%s" % output_path, filepath]
                 )
     except FileNotFoundError:
-        utils.if_callable_call_with_formatted_string(status_callback, GS_NOT_INSTALLED,
-                                                     ghostscript_binary)
+        if_callable_call_with_formatted_string(status_callback, GS_NOT_INSTALLED,
+                                               ghostscript_binary)
         sys.exit(1)
     process.communicate()
-    utils.if_callable_call_with_formatted_string(status_callback, FILE_DONE, output_path)
+    if_callable_call_with_formatted_string(status_callback, FILE_DONE, output_path)
 
 def compress_multiple_pdfs(source_directory, output_directory, ghostscript_binary, status_callback=None):
     """Compress all PDF files in the current directory and place the output in the given output directory.
@@ -97,11 +97,11 @@ def compress_multiple_pdfs(source_directory, output_directory, ghostscript_binar
     """
     source_paths = get_pdf_filenames_at(source_directory)
     out_paths = list()
-    utils.if_callable_call_with_formatted_string(status_callback, COMPRESSING_MULTIPLE,
-                                                 source_directory, output_directory, len(source_paths))
+    if_callable_call_with_formatted_string(status_callback, COMPRESSING_MULTIPLE,
+                                           source_directory, output_directory, len(source_paths))
     for source_path in source_paths:
         output = os.path.join(output_directory, os.path.basename(source_path))
         out_paths.append(output)
         compress_pdf(source_path, output, ghostscript_binary, status_callback)
-    utils.if_callable_call_with_formatted_string(status_callback, ALL_FILES_DONE, output_directory)
+    if_callable_call_with_formatted_string(status_callback, ALL_FILES_DONE, output_directory)
     return out_paths
