+++ tests/context.py
@@ -2,4 +2,7 @@
 import sys
 import os
 sys.path.insert(0, os.path.abspath(os.path.join(os.path.dirname(__file__), '..')))
-import pdfebc_core.compress, pdfebc_core.utils
+import pdfebc_core.compress
+import pdfebc_core.misc_utils
+import pdfebc_core.email_utils
+import pdfebc_core.config_utils
